fisrt commit

+++ src/main/java/com/aylfq5/online/tutor/dao/StudentInfoMapper.java
@@ -3,13 +3,13 @@ package com.aylfq5.online.tutor.dao;
 import com.aylfq5.online.tutor.domain.StudentInfo;
 
 public interface StudentInfoMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(StudentInfo record);
 
     int insertSelective(StudentInfo record);
 
-    StudentInfo selectByPrimaryKey(Integer id);
+    StudentInfo selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(StudentInfo record);
