+++ src/main/java/com/aylfq5/online/tutor/dao/StudentVolunteerMapper.java
@@ -3,13 +3,13 @@ package com.aylfq5.online.tutor.dao;
 import com.aylfq5.online.tutor.domain.StudentVolunteer;
 
 public interface StudentVolunteerMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(StudentVolunteer record);
 
     int insertSelective(StudentVolunteer record);
 
-    StudentVolunteer selectByPrimaryKey(Integer id);
+    StudentVolunteer selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(StudentVolunteer record);
