+++ src/main/java/com/aylfq5/online/tutor/dao/UserMapper.java
@@ -3,13 +3,13 @@ package com.aylfq5.online.tutor.dao;
 import com.aylfq5.online.tutor.domain.User;
 
 public interface UserMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(User record);
 
     int insertSelective(User record);
 
-    User selectByPrimaryKey(Integer id);
+    User selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(User record);
